requirements.txt
# local package
-e .

# external packages
jupyter
ipykernel
pandas
matplotlib
openai
python-dotenv
chromadb